text-labeling-assistant/requirements.txt
streamlit
pandas
scikit-learn
joblib
transformers
torch
sentencepiece
datasets
openai
python-dotenv
